Sesuaikan nama file dengan unit test; tambahkan unit testing untuk pembacaan tabel dan kolom
